made userid an int, which ended up touching a lot
Former-commit-id: de3d509ad3538ecb419fc30f364e37090a5af668
Former-commit-id: 64b30412a7a867a4996a51baeeb82ce9eb33b40d

+++ src/mitll/langtest/shared/exercise/CommonExercise.java
@@ -37,7 +37,7 @@ public interface CommonExercise extends CommonAudioExercise, ScoredExercise {
    * @see mitll.langtest.client.custom.dialog.EditItem#didICreateThisItem(CommonExercise)
    * @return
    */
-  long getCreator();
+  int getCreator();
 
   long getUpdateTime();
